Added conditions for Talk to Self showing up as a command.
The following will stop the command from showing up:
- Character is in fight mode
- A boarding is in progress
- Player alarm is active (character is in a fight)
- The player is not currently controlling the character
- The character is on Deck

If the Character enters the deck with the "BI_EnterShip" interface command and then tries to talk to themself, the game crashes. That last condition is a temporary fix for this.

I honestly have no idea why the crash occurs. I'm open to suggestions.